requirements.txt
apache-airflow
requests
beautifulsoup4
mysql-connector-python